+++ lib/absinthe/lexer.ex
@@ -275,13 +275,13 @@ defmodule Absinthe.Lexer do
     )
   )
 
-  defp fill_mantissa(_rest, raw, context, _, _), do: {'0.' ++ raw, context}
+  defp fill_mantissa(rest, raw, context, _, _), do: {rest, '0.' ++ raw, context}
 
-  defp unescape_unicode(_rest, content, context, _loc, _) do
+  defp unescape_unicode(rest, content, context, _loc, _) do
     code = content |> Enum.reverse()
     value = :erlang.list_to_integer(code, 16)
     binary = :unicode.characters_to_binary([value])
-    {[binary], context}
+    {rest, [binary], context}
   end
 
   @boolean_words ~w(
@@ -315,48 +315,50 @@ defmodule Absinthe.Lexer do
     do_boolean_value_or_name_or_reserved_word(rest, value, context, loc, byte_offset)
   end
 
-  defp do_boolean_value_or_name_or_reserved_word(_rest, value, context, loc, byte_offset)
+  defp do_boolean_value_or_name_or_reserved_word(rest, value, context, loc, byte_offset)
        when value in @boolean_words do
-    {[{:boolean_value, line_and_column(loc, byte_offset, length(value)), value}], context}
+    {rest, [{:boolean_value, line_and_column(loc, byte_offset, length(value)), value}], context}
   end
 
-  defp do_boolean_value_or_name_or_reserved_word(_rest, value, context, loc, byte_offset)
+  defp do_boolean_value_or_name_or_reserved_word(rest, value, context, loc, byte_offset)
        when value in @reserved_words do
     token_name = value |> List.to_atom()
-    {[{token_name, line_and_column(loc, byte_offset, length(value))}], context}
+    {rest, [{token_name, line_and_column(loc, byte_offset, length(value))}], context}
   end
 
-  defp do_boolean_value_or_name_or_reserved_word(_rest, value, context, loc, byte_offset) do
-    {[{:name, line_and_column(loc, byte_offset, length(value)), value}], context}
+  defp do_boolean_value_or_name_or_reserved_word(rest, value, context, loc, byte_offset) do
+    {rest, [{:name, line_and_column(loc, byte_offset, length(value)), value}], context}
   end
 
-  defp labeled_token(_rest, chars, context, loc, byte_offset, token_name) do
+  defp labeled_token(rest, chars, context, loc, byte_offset, token_name) do
     value = chars |> Enum.reverse()
-    {[{token_name, line_and_column(loc, byte_offset, length(value)), value}], context}
+    {rest, [{token_name, line_and_column(loc, byte_offset, length(value)), value}], context}
   end
 
-  defp mark_string_start(_rest, chars, context, loc, byte_offset) do
-    {[chars], Map.put(context, :token_location, line_and_column(loc, byte_offset, 1))}
+  defp mark_string_start(rest, chars, context, loc, byte_offset) do
+    {rest, [chars], Map.put(context, :token_location, line_and_column(loc, byte_offset, 1))}
   end
 
-  defp mark_block_string_start(_rest, _chars, context, loc, byte_offset) do
-    {[], Map.put(context, :token_location, line_and_column(loc, byte_offset, 3))}
+  defp mark_block_string_start(rest, _chars, context, loc, byte_offset) do
+    {rest, [], Map.put(context, :token_location, line_and_column(loc, byte_offset, 3))}
   end
 
-  defp block_string_value_token(_rest, chars, context, _loc, _byte_offset) do
+  defp block_string_value_token(rest, chars, context, _loc, _byte_offset) do
     value = '"""' ++ (chars |> Enum.reverse()) ++ '"""'
-    {[{:block_string_value, context.token_location, value}], Map.delete(context, :token_location)}
+
+    {rest, [{:block_string_value, context.token_location, value}],
+     Map.delete(context, :token_location)}
   end
 
-  defp string_value_token(_rest, chars, context, _loc, _byte_offset) do
+  defp string_value_token(rest, chars, context, _loc, _byte_offset) do
     value = '"' ++ tl(chars |> Enum.reverse()) ++ '"'
-    {[{:string_value, context.token_location, value}], Map.delete(context, :token_location)}
+    {rest, [{:string_value, context.token_location, value}], Map.delete(context, :token_location)}
   end
 
-  defp atom_token(_rest, chars, context, loc, byte_offset) do
+  defp atom_token(rest, chars, context, loc, byte_offset) do
     value = chars |> Enum.reverse()
     token_atom = value |> List.to_atom()
-    {[{token_atom, line_and_column(loc, byte_offset, length(value))}], context}
+    {rest, [{token_atom, line_and_column(loc, byte_offset, length(value))}], context}
   end
 
   def line_and_column({line, line_offset}, byte_offset, column_correction) do