+++ lib/absinthe/phase/document/result.ex
@@ -61,9 +61,7 @@ defmodule Absinthe.Phase.Document.Result do
                 Could not serialize term #{inspect(value)} as type #{schema_node.name}
 
                 When serializing the field:
-                #{emitter.parent_type.name}.#{emitter.schema_node.name} (#{
-                  emitter.schema_node.__reference__.location.file
-                }:#{emitter.schema_node.__reference__.location.line})
+                #{emitter.parent_type.name}.#{emitter.schema_node.name} (#{emitter.schema_node.__reference__.location.file}:#{emitter.schema_node.__reference__.location.line})
                 """
               )
           end